refactor 역할에 맞도록 함수 이동, 과도한 인터페이스 제거, 테스트를 위한 패키지 이동
