+++ src/test/java/domain/SameNumberGenerator.java
